实现背景图片的选取。

+++ JiaMian/CustomClass/NetWorkConnect.h
@@ -33,7 +33,7 @@
 - (MessageModel*)messageLikeByMsgId:(long)MsgId;
 
 //MsgType: 1; AreaId: 1; lat: 0.0; lom: 0.0
-- (MessageModel*)messageCreate:(NSString*)Text msgType:(int)MsgType areaId:(int)AreaId lat:(double)Lat lon:(double)Long;
+- (MessageModel*)messageCreate:(NSString*)Text msgType:(int)MsgType areaId:(long)AreaId lat:(double)Lat lon:(double)Long;
 
 //SinceId: 0; MaxId: INT_MAX; Count: 20
 - (NSArray*)commentShowByMsgId:(long)MsgId sinceId:(long)SinceId maxId:(long)MaxId count:(int)Count;
